Truncate all the float literals
Float literals don't behave intuitively under FLT_EVAL_METHOD == 2.

+++ tests.c
@@ -115,7 +115,7 @@ main(int argc, char **argv)
     TEST("size init", buf_size(a) == 0);
     buf_push(a, 1.3f);
     TEST("size 1", buf_size(a) == 1);
-    TEST("value", a[0] == 1.3f);
+    TEST("value", a[0] == (float)1.3f);
     buf_free(a);
     TEST("free", a == 0);
 
@@ -144,12 +144,12 @@ main(int argc, char **argv)
     buf_push(a, 1.3);
     buf_push(a, 1.4);
     TEST("size 4", buf_size(a) == 4);
-    TEST("pop 3", buf_pop(a) == 1.4f);
+    TEST("pop 3", buf_pop(a) == (float)1.4f);
     buf_trunc(a, 3);
     TEST("size 3", buf_size(a) == 3);
-    TEST("pop 2", buf_pop(a) == 1.3f);
-    TEST("pop 1", buf_pop(a) == 1.2f);
-    TEST("pop 0", buf_pop(a) == 1.1f);
+    TEST("pop 2", buf_pop(a) == (float)1.3f);
+    TEST("pop 1", buf_pop(a) == (float)1.2f);
+    TEST("pop 0", buf_pop(a) == (float)1.1f);
     TEST("size 0 (pop)", buf_size(a) == 0);
     buf_free(a);
